File-based config system
Summary:
The Python module thing with a command-line arg was nice, but broke with the various scripts and alembic.

These config files won't be changing often, so I feel OK putting them in a file on disk.

Obviously the hardcoded paths can in the future be pulled from environment variables or something, but this works for now and is important to move forward.

Test Plan: Tests pass. You need a test config

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D177