Fix IntegrityError when syncing a new folder.
Previously, we'd call get_associated_tag from Folder.find_or_create to get or
create the tag associated to a new folder. If we then add that folder to a
thread's folders, the also_set_tag validator would query for that tag. Because
we have autoflush disabled while constructing threads in the mailsync code,
this query wouldn't find the tag. So we'd create it again. On the next
flush, this caused an IntegrityError.