Get crispin client inside code wrapped by retry_crispin, not outside.
Summary:
We were passing a crispin_client instance to functions wrapped by
`@retry_crispin`. That meant that if Gmail closed the connection, we'd be
retrying with the *same* connection, leading to inevitable failure.

Instead, pass the connection pool, and get a new client from within the wrapped
code. This way we'll use a new connection on retry.

Test Plan:
I ran side-by-side syncs with and without this change. Both
eventually logged a "socket error: EOF". The sync with this change successfully
retried with a new connection, the sync without retried several times before
aborting.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D181