Sync modified gmail labels on existing, fully-synced threads.
Summary:
If we had already fully synced a thread, and you added or removed a label on it
in the Gmail web interface, we wouldn't persist that modification when we
called update_metadata(). This change fixes that. It also simplifies some logic
dealing with label name case-sensitivity, because we're now exposing them in
lowercase anyways.

Test Plan: Manual sync testing.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D180