[DEPENDENCIES] Don't install supervisor from PyPI.
This just doesn't seem to work on Debian/Ubuntu; use the version from
APT instead, which we already install in setup.sh. If you are attempting
to run a proper prod setup, you will wish to 'pip uninstall supervisor'.
(If this doesn't work, you'll have to manually remove the files in
/usr/local.)