Don't call drop_everything() when tearing down the test db fixture.
Summary:
This call is failing with an odd sqlalchemy ResourceClosedError in certain
configurations of some tests for the drafts API that I'm working on.
I really don't know why. But in any case we don't actually need it since we
load the database dump from scratch for each test. In fact, eliminating it
makes the tests run a few seconds faster.

Test Plan: py.test tests

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D188