Fix bug preventing EAS accounts from being authed.
Commit 4cefd27 removed a few lines from inbox.auth.handler_from_provider that
were needed to auth EAS accounts. Basically, if the provider wasn't Gmail or
Yahoo, we'd always try EAS. This commit does the same thing in
inbox.util.url.provider_from_address. This strategy isn't ideal in the longer
term, but is better than not being able to auth an EAS account at all.)

(Note that nothing changes if EAS support is not installed.)