Fix KeyError in Gmail mailsync's add_new_imapuids()
Messages can still move around while a given IMAP session is open, so
sometimes we'll start adding ImapUid entries but not get the flag info
while the message is still in the given folder. (For example, when a
message is archived in the middle of this operation.)

The right thing to do in this situation is to skip adding the ImapUid
entry; we'll catch the message later via the All Mail sync.