Fixes T194: IntegrityError (1452) FolderItem.folder_id -> Folder.id Foreign Key Constraint Fails.
Summary:
Here's what I think the issue is:

We create a new Folder (line 118, mailsync.base.create_db_objects()) but do not add it to the session.
The corresponding FolderItems are created in line 680, mailsync.gmail.add_gmail_attrs() with autoflush disabled.
Then in contacts.process_mail.update_contacts(), we call get_contact_objects();
the db_session.query(Contact).filter(…).all() in this function triggers an autoflush and causes
the IntegrityError (1452): foreign key constraint fails for FolderItem.folder_id -> Folder.id because the FolderItems are in the session but the Folder
is not.

To fix this:
Added the Folder to the session in mailsync.base.create_db_objects()

Note:
I am not certain that this is the issue, I could not repro it to test the fix - hence, the request for review.

Test Plan: Sync runs as before.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D263