[auth/unknown] Better handling of unknown providers.
Summary:
If we don't know what the provider is, we can still try to auth through
exchange (though don't show exchange logo for this) and if that fails
then gracefully tell the user that we don't support that provider.

Support for gathering emails at that point should be added at the front-end.

Refs: T213

Test Plan: review code.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D270