Explicitly garbage collect ScriptDirectory object in check_db().
For whatever reason, the line `script.get_current_head()` leaves a bunch
of idle db connections lying around until the ScriptDirectory object is
garbage collected. I'm not sure what it's using them all for because
ahhh source diving SQLAlchemy, but it's sure not because we're asking
for a lot of connections in our migration env config. For now, we can
work around by explicitly cleaning them up.