Configurable resettable_counter.
Summary:
retry_crispin sets max_count=5, reset_interval=150 in the hope
this decreases the number of imaplib abort errors we've been seeing.

We pick 5 because that was the number we'd used with geventconnpool's retry
(i.e. before we switched to our own). 150 is more arbitrary.

Test Plan: Test added to tests/general/test_concurrency.py

Reviewers: spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D286