[DEPENDENCIES] Add facility to cumulatively profile sync processes.
Summary:
This commit adds the ability to get cumulative profiler data for a running sync
process by doing `kill -SIGTRAP <sync_pid>`, if the 'DEBUG_PROFILING_ON' flag
is set in the config.

Requires an updated version of pyinstrument to avoid issues around interrupting
syscalls when the profiler runs with use_signal=True.

Test Plan: run a mailsync and try it out.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D303