Generic Condstore syncing, take 2.
Summary:
Christine, Charles,

I extracted the condstore syncing code from D280 and adapted it to the new generic backend.

The code is functional but there's some additional work to do which is out of the scope of this patch.

1. There's some refactoring to do around "crispin_new". Mostly, I'd like to sniff out the capabilities of the server before constructing the CrispinClient. Right now, every non-gmail provider gets a GenericCrispinClient.

2. There's still this bug in ImapClient. Menno told me he was real close to releasing a fix. Should I go a ahead and fork ImapClient temporarily or just wait for the release?

Test Plan: Ran a couple syncs with fastmail + checked in the ImapClient logs that it was issuing the condstore commands.

Reviewers: charles, spang

Reviewed By: spang

Differential Revision: https://review.inboxapp.com/D297