Fix Sentry message truncation.
In an illustrative example of the hazards of constructing an untyped data model
out of dictionaries and lists, raven's RemoveStackLocalsProcessor doesn't do
what it says on the box, so neither does our code based on it. Put together
something that should work better until we can submit a PR upstream.