Fix MX domain validation 'bug'
Summary: I spent an hour on this quasi-bug... *sigh* The problem is the MX domain validation code expects domain names to end with a '.' (because this is how they are represented in DNS). I changed the code to validate against standard domain names.

Test Plan: Ran a sync on gandi.

Reviewers: charles

Reviewed By: charles

Differential Revision: https://review.inboxapp.com/D318