[default_folder_map] Add support for using default folder mappings and looking at flags on folders to determine what system folders they are.
Summary:
Previously we had a folder map for each provider, but many of
them shere common names. Additionally, some providers use flags on
certain folders do describe what they are. This is particularly useful
in the situation where the actual folder names might be in a different
language. Use these flags and a default folder map to clean up the
providers.py file so that it only needs to special-case certain folders
that don't fit into the default mapping.

Test Plan: Review the code.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D348