Don't use an SMTPConnectionPool.
Summary:
We don't really need to maintain an SMTP connection pool, because in general we
only rarely need an SMTP connection for an account. Currently, once we open a
connection we try to keep it around forever, which is not ideal.

Instead, we just open and close a connection on an as-needed basis. This
resolves issues we were seeing around messages not being sent because of
improperly handled SMTPServerDisconnected errors.

Test Plan:
Use system tests to send messages for various providers; use added
raise_once utility in inbox/util/debug.py to check that sending succeeds when
retrying after an SMTPServerDisconnected error.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D357