Create archive folder when needed
Summary: It seems someone tried to archive emails for a non-gmail provider on Beta. This resulted into a large number of errors (https://app.getsentry.com/inbox/beta/group/29293091/). This diff fixes this bug (T328).

Test Plan: Added an archived tag to a thread in yahoo. Checked that the folder was created.

Reviewers: spang, emfree

Reviewed By: emfree

Maniphest Tasks: T328

Differential Revision: https://review.inboxapp.com/D367