Wrap greenlet-spawning code inside a try/finally block
Summary:
Eben,

This is a patch extracted from the account deletion diff. It simply wraps the initial sync code inside a try/finally block so that the greenlets we spawn during the initial sync get killed when the main greenlet is killed.

Test Plan: Ran a sync.

Reviewers: emfree

Reviewed By: emfree

Maniphest Tasks: T330

Differential Revision: https://review.inboxapp.com/D380