[api/tags] Accept the namespace_id parameter for tag creation and updates. Add tests for better coverage of /tags api endpoint.
Summary:
The /tags endpoint previously would not accept a "namespace_id" parameter, even though it returns one when you create it. This not only doesn't agree with RESTful semantics (you should be able to replace the object in-full using a PUT / POST), it also doesn't gel with other endpoints and requires some work-arounds in the python libraries to special case tags. Instead we now accept the 'namespace_id' parameter as well as tests for full coverage of the /tags endpoint.

Fixes T359

Test Plan: run the tests.

Reviewers: spang

Reviewed By: spang

Maniphest Tasks: T359

Differential Revision: https://review.inboxapp.com/D392