[build/setup.sh] Add libsodium setup to install flow. NOTE: for some reason we hang on pip install -e .
Summary:
We now require libsodium as a dependency, however it wasn't a part of the install flow. This patch adds it to the install flow. Since there isn't an ubuntu package available we download (check hash) and install from source. As noted in the title, the setup for some reasn seems to be hanging on 'pip install -e .' for me. If that happens for others we should file a task, otherwise I can ignore.

Fixes T366

Test Plan: checkout master without libsodium installed, run the setup and verify that you can install the pynacl package.

Reviewers: kav-ya, spang, dlitz

Reviewed By: spang, dlitz

Subscribers: dlitz

Projects: #security

Maniphest Tasks: T366

Differential Revision: https://review.inboxapp.com/D393